--- watch/src/blockprint/error.rs
@@ -1,1 +1,0 @@
-
