+++ watch/src/lib.rs
@@ -1,3 +1,5 @@
+pub mod block_packing;
+pub mod block_rewards;
 pub mod blockprint;
 pub mod cli;
 pub mod client;
@@ -5,4 +7,5 @@ pub mod config;
 pub mod database;
 pub mod logger;
 pub mod server;
+pub mod suboptimal_attestations;
 pub mod updater;
